Fix pip installation error when Cython is missing
Installing hidapi through pip fails when Cython is not installed
because the import of Cython.Distutils fails.  This patch catches
the exception when importing Cython.Distutils and ignores it.  This
allows setup to complete, even without having Cython already installed.
After the initial installation of hidapi the command build_ext can be
used.