Add --without-libusb option to setup.py
This option allows to build and install cython-hidapi
with hidraw support only on Linux. Thus libusb dependency
becomes optional.

Because there is no standard support in setuptools for
--with-<feature>/--without-<feature> options during building
we choose the simplest solution suggest at [1].

Also we change the hidraw extension module name from 'hidraw'
to 'hid' in case --without-libusb is specified.
This is required to keep external cython-hidapi API consistent
and not to bother user with backend chosen.

[1] - http://stackoverflow.com/questions/677577/distutils-how-to-pass-a-user-defined-parameter-to-setup-py