Remove hidapi_exit() entirely
It was not supposed to be used explicitly in the first place (see
comments on the PR this commit is a part of).  The HIDAPI library will
still be cleanly finalized before the program terminates, through
another – automatic – mechanism.

Additionally, the only currently known caller is using it as a
workaround for another issue (#114), and that use case would not be
supported by the no-op compatibility shim anyway.

Related: #114 ("Cannot reenumarate devices in same python process")
Related: #128 ("Fatal Python error: Segmentation fault on 0.11.0.post2")