Change the default backend on Linux to HIDRAW
While there are tradeoffs to using either the backend, HIDRAW is thought
to be superior to LibUSB except in a very few cases.[^1][^2]

Thus, only build the HIDRAW backend by default, and name the resulting
module hid.  This effectively makes the old --without-libusb switch the
default behavior.

Additionally, add a --with-libusb switch for the LibUSB backend to be
built as well.  In that case, and to avoid breaking exiting code that
already dynamically picks its preferred backend, the resulting module
names are kept the same as before – hid for LibUSB and hidraw for HIDRAW
– and LibUSB will be used by default when simply importing hid.

[^1]: https://github.com/trezor/cython-hidapi/issues/110#issuecomment-779188978
[^2]: https://github.com/libusb/hidapi/tree/4ab60e005eb5#linuxhidraw-linuxhidc