Remove references to self when creating device finalizers
From the documentation:

> It is important to ensure that func, args and kwargs do not own any
references to obj, either directly or indirectly, since otherwise obj
will never be garbage collected. In particular, func should not be a
bound method of obj.

The issue wasn't observed during testing, but it may be
implementation-defined (as with __del__) and depend on specific
combinations of interpreter and Cython versions.

Instead of relying on unspecified behavior, only keep the hid_device
pointer.  This requires a wrapper because weakref.finalize() cannot be
passed cdef functions or C pointers.

Alternatives were attempted---__dealloc__, weakref.ref(self)---but
they only worked in some cases (respectively: only for `del <device>`
and except for `del <device>`).

Reported-by: @bearsh
Related: #130