expose hid_version_str
`hid_version_str` was added to the underlying C library in 0.10.0,
in https://github.com/libusb/hidapi/pull/192

Note that the method added here returns the version of the C library,
which is different from the version of cython-hidapi, i.e. the `hid` python package.
It would be good to also make the latter available at runtime somehow.